frontend/requirements.txt
aiohttp==3.8.6
discord.py==2.3.2
Flask==3.0.0
gunicorn==20.1.0
mysql.connector
pytz==2023.3.post1
Requests==2.31.0
websockets==11.0.3
Zenora==0.0.3.post1

